requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
bidict==0.23.1
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
docutils==0.22.4
fastapi==0.135.1
Flask==3.1.3
fonttools==4.62.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.5.0
lxml==6.0.2
markdown2==2.5.5
MarkupSafe==3.0.3
matplotlib==3.10.8
msal==1.35.1
multidict==6.7.1
nicegui==3.8.0
numpy==2.4.3
oauthlib==3.3.1
orjson==3.11.7
packaging==26.0
pefile==2024.8.26
pillow==12.1.1
propcache==0.4.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
PyJWT==2.12.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-engineio==4.13.1
python-multipart==0.0.22
python-socketio==5.16.1
pywin32-ctypes==0.2.3
PyYAML==6.0.3
qrcode==8.2
requests==2.32.5
requests-oauthlib==2.0.0
setuptools==82.0.1
simple-websocket==1.1.0
six==1.17.0
soupsieve==2.8.3
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.6
wsproto==1.3.2
yarl==1.23.0
